Home_Work_25_10/requirements.txt
# requirements.txt
selenium==4.37.0
pytest==8.4.2
allure-pytest==2.15.0
questionary==2.1.1